Arreglos Iteracion 1

+++ src/main/java/uniandes/isis2304/b07/superandes/negocio/VOProducto.java
@@ -12,7 +12,7 @@ public interface VOProducto {
 	
 	public String getPresentacion();
 	
-	public int getCantidad();
+	public long getCantidad();
 	
 	public String getUnidadMedida();
